no longer failing tests and also good lint times
